+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## Unreleased
 
-
 ### Features
 
 - dbaas: add missing PostgreSQL configuration options (`shared-buffers-percentage`, `synchronous-replication`, `timescaledb-settings`, `variant`, `work-mem`) #808
 - `exo compute eip list` now shows the description & the management type of the EIP #803
 - sks: add `rotate-karpenter-credentials` command #797
 - sks: add `active-nodepool-templates` command #797
+- new command `exo ai deployment instance-type` that allows showing what GPU is usable in which zone #809
 
 ### Bug fixes